study 230807
bug fixed

+++ src/controllers/videoController.js
@@ -30,9 +30,13 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const { user: { _id }, } = req.session
   const video = await Video.findById(id);
   if (!video) {
     return res.status(400).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/")
   }
   return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
@@ -40,10 +44,14 @@ export const getEdit = async (req, res) => {
 // 같은 render는 한번만 가능
 export const postEdit = async (req, res) => {
   const { id } = req.params;
+  const { user: { _id }, } = req.session
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
   if (!video) {
     return res.render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/")
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -102,6 +110,14 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const { user: { _id }, } = req.session
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(400).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/")
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
